requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
automium==0.2.6
automium-web==0.1.1
autopep8==1.4.4
beautifulsoup4==4.9.3
certifi==2019.3.9
cffi==1.14.6
chardet==3.0.4
click==7.1.2
colorama==0.4.4
comtypes==1.1.8
cryptography==3.4.8
decorator==5.1.0
defusedxml==0.6.0
distlib==0.3.1
Django==2.2.14
django-allauth==0.39.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-filter==2.4.0
django-rest-knox==4.1.0
django-rest-multiple-models==2.1.3
djangorestframework==3.12.4
dynaconf==3.1.7
filelock==3.0.12
Flask==0.12.4
future==0.18.2
hvac==0.11.2
idna==2.8
image==1.5.33
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
knox==0.1.14
lazy-object-proxy==1.4.3
legacy==0.1.6
loguru==0.5.3
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
MouseInfo==0.1.3
Naked==0.1.31
oauthlib==3.0.1
pefile==2019.4.18
pep8==1.7.1
Pillow==8.1.0
PyAutoGUI==0.9.52
pycodestyle==2.5.0
pycparser==2.20
PyGetWindow==0.0.9
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pyjokes==0.6.0
pylint==2.6.0
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.2
PyQt5-sip==12.8.1
PyQtWebEngine==5.15.2
PyRect==0.1.4
PyScreeze==0.1.26
python-decouple==3.1
python3-openid==3.1.0
pyttsx3==2.90
PyTweening==1.0.3
pytz==2018.5
pywhatkit==3.9
pywin32==300
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.21.0
requests-oauthlib==1.2.0
selenium==3.141.0
Serializer==0.2.1
serializers==0.2.4
shortuuid==1.0.1
simplejson==3.17.5
six==1.15.0
soupsieve==2.2
SpeechRecognition==3.8.1
sqlparse==0.2.4
status==0.2.5
stripe==2.27.0
toml==0.10.2
urllib3==1.24.2
users==1.0.dev0
validators==0.18.2
views==0.3
virtualenv==20.2.2
Werkzeug==0.12.2
wikipedia==1.4.0
win32-setctime==1.0.3
wrapt==1.12.1
